Split ts remover
